pyproject.toml
# Poetry build chain
[tool.poetry]
name = "aisplanner"
version = "0.1.0"
description = ""
authors = ["Niklas Paulig <niklas-paulig@tu-dresden.de>"]
readme = "README.md"
packages = [{include = "aisplanner"}]

[tool.poetry.dependencies]
python = "^3.7"
numpy = "^1.17.2"
matplotlib = "^3.1.3"
pandas = "^1.0.0"
pyais = "^2.0.0"
sklearn = "*"
paramiko = "^2.8.0"
python-dotenv = "^1.0.0"
ciso8601 = "^2.3.0"
utm = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
